+++ src/index.ts
@@ -159,6 +159,13 @@ export const options = {
     description:
       "Keep the open and closing brackets on the same line for the empty blocks.",
   },
+  apexExpandOneLineProperties: {
+    type: "boolean",
+    category: CATEGORY_APEX,
+    default: true,
+    description:
+      "Expand properties even if there is only one statement in getter or setter. If the option is `false` the properties will be inline if there is only one statement or if the `printWidth` option is not exceeded.",
+  },
 };
 
 export const defaultOptions = {};